test: 프로덕션 코드에 맞춰 테스트 클래스 통합
